[BTAT-116] Refactored the MicroserviceHelloWorld controller to be the starting point for the EstimatedTaxLiabilityController
