+++ app/controllers/EstimatedTaxLiabilityController.scala
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2017 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package controllers
+
+import javax.inject.{Inject, Singleton}
+
+import config.AppConfig
+import controllers.predicates.AuthenticationPredicate
+import play.api.mvc._
+import uk.gov.hmrc.play.microservice.controller.BaseController
+
+import scala.concurrent.Future
+
+@Singleton
+class EstimatedTaxLiabilityController @Inject()(implicit val appConfig: AppConfig,
+                                                val authentication: AuthenticationPredicate
+                                      ) extends BaseController {
+
+  val getEstimatedTaxLiability: Action[AnyContent] = authentication.async { implicit request =>
+		Future.successful(Ok("Hello world"))
+  }
+}
