+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             @controllers.MicroserviceHelloWorld.hello
+GET        /estimated-tax-liability             @controllers.EstimatedTaxLiabilityController.getEstimatedTaxLiability
