+++ conf/prod.routes
@@ -1,5 +1,5 @@
 # Add all the application routes to the app.routes file
 ->         /income-tax-view-change              app.Routes
-->         /                          health.Routes
+->         /                                    health.Routes
 
-GET        /admin/metrics             @com.kenshoo.play.metrics.MetricsController.metrics
+GET        /admin/metrics                       @com.kenshoo.play.metrics.MetricsController.metrics
